Add files via upload
Based on the original Demo.csv file, which has been cleaned up by taken out the blanks rows and considering that the are 2 classification of cases, so in the original data set, when you sum up all cases the total is greater than what is shown in the third column. I am also submitting another couple of cleaned files for Illinois data only.